pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "gs1-digital-link-resolver"
version = "0.1.0"
description = "Open GS1 Digital Link Resolver for EU Digital Product Passports"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
keywords = ["gs1", "digital-link", "dpp", "espr", "supply-chain", "resolver"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[project.urls]
Homepage = "https://github.com/grigolatoe/gs1-digital-link-resolver"
"Bug Tracker" = "https://github.com/grigolatoe/gs1-digital-link-resolver/issues"
